backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
mysqlclient==2.2.0
python-dotenv==1.0.0
django-filter==23.5
Pillow==10.0.0
djangorestframework-simplejwt[crypto]==5.3.0
PyJWT==2.8.0
social-auth-app-django==5.4.0
google-auth==2.23.0
requests==2.31.0 